Retrieves Game Details Externally
For the most part, I was copying and pasting details from the website "How Long To Beat".

I began to wonder if I could engineer a way to have the website retrieve those details for itself based on an ID I provided.

I created a hook and an API endpoint and reworked a lot of the structure to account for the now missing "Details" section which is no longer necessary.

Unfortunately, this has resulted in removing some columns/metadata from the lists, as well.

But given that this change will _significantly_ increase the speed at which I can add games, I am fine with that trade.